Learned how to add images, image maps, style horizontal rules, and a background image: Chapters 8 through 12.
I hope that I may get the setup on GitHub as it is on my local machine.  
1.  tpa_html is the lead sheet, it calls the other html sheets when the given image of a Saturn Moon is clicked.
2.  the images folder is the one all sheets look to to populate the page with an image.

====================

Okay:  I cannot figure out how to load a folder; I must change the code in the sheets and remove the images files from their folder and put them with the sheet files here on GitHub.  The "G" after the htmt names is to identify them separate from the original work.